Agregando swal

+++ pwaTis/PWA/src/app/services/score.service.ts
@@ -6,6 +6,8 @@ import { Rating, Score, ScoreIDB } from '../models/score.interface';
 import Dexie from 'dexie';
 import { CheckStatusService } from './check-status.service';
 import { OnlineStatusService, OnlineStatusType } from 'ngx-online-status';
+import Swal from 'sweetalert2'
+
 
 
 
@@ -24,7 +26,7 @@ export class ScoreService {
     status: OnlineStatusType; //Enum provided by ngx-online-status
     offline: any;
 
-    constructor(private onlineStatusService: OnlineStatusService,private checkStatusService: CheckStatusService, private http: HttpClient) {
+    constructor(private onlineStatusService: OnlineStatusService, private checkStatusService: CheckStatusService, private http: HttpClient) {
 
 
         this.startIndexedDB();
@@ -33,17 +35,42 @@ export class ScoreService {
             // Retrieve Online status Type
             this.status = status;
             this.offline = (status === this.onlineStatusCheck.OFFLINE)
-            
-            if(!this.offline){
-                console.log("ONLINE");
+
+            if (!this.offline) {
+                Swal.fire({
+                    title: 'Reconectado',
+                    icon: 'success',
+                    confirmButtonText: 'Aceptar',
+                    width: '20%',
+                    backdrop: false,
+                    timer: 3000,
+                    toast: true,
+                    position: 'top-end'
+                }).then(
+                    response =>{
+                        
                 this.sendIndexedToApi();
+                    }
+                )
             }
-            
-            
-           
-            
-            
-          });
+            else{
+                Swal.fire({
+                    title: 'Desconectado',
+                    icon: 'warning',
+                    confirmButtonText: 'Aceptar',
+                    width: '20%',
+                    backdrop: false,
+                    timer: 3000,
+                    toast: true,
+                    position: 'top-end'
+                })
+            }
+
+
+
+
+
+        });
 
     }
 
@@ -57,6 +84,17 @@ export class ScoreService {
 
     public async saveRatingIndexDB(rating: ScoreIDB) {
 
+        Swal.fire({
+            title: 'Sin conexión a internet, guardaremos su calificación localmente hasta que se reanude la conexión',
+            icon: 'warning',
+            confirmButtonText: 'Aceptar',
+            width: '20%',
+            backdrop: false,
+            timer: 3000,
+            toast: true,
+            position: 'top-end'
+        })
+
         await this.table.add(rating);
         const todosRatings: ScoreIDB[] = await this.table.toArray();
         console.log('Ratings ', todosRatings);
@@ -68,29 +106,49 @@ export class ScoreService {
 
         const todosRatings: ScoreIDB[] = await this.table.toArray()
 
-        console.log('Ratings ', todosRatings);
+        if (todosRatings.length > 0) {
 
+            Swal.fire({
+                title: 'Conexión retomada, sincronizaremos su información con el servidor',
+                icon: 'success',
+                confirmButtonText: 'Aceptar',
+                width: '20%',
+                backdrop: false,
+                timer: 3000,
+                toast: true,
+                position: 'top-end'
+            }).then(
+                response => {
+                    for (const rating of todosRatings) {
+                        this.mergeScore(rating).subscribe(
+                            (async response => {
+                                await this.table.delete(rating.id)
+                            }
+                            )
+                        );
 
-        for (const rating of todosRatings) {
-            this.mergeScore(rating).subscribe(
-                (async response => {
-                    await this.table.delete(rating.id)
+                    }
                 }
-                )
-            );
+            )
+
+
+
+
+
 
         }
+        console.log();
+
+
+
     }
 
     public mergeScore(score: ScoreIDB): Observable<any> {
 
-        if (this.checkStatusService.status === this.onlineStatusCheck.OFFLINE) {
+        if (this.offline) {
             this.saveRatingIndexDB(score)
         }
         else {
-
-            console.log(score);
-
 
             let newScore: Rating = {
                 scoreId: {
@@ -103,7 +161,19 @@ export class ScoreService {
                 },
                 scoreNumber: score.scoreNumber
             }
+
+            Swal.fire({
+                title: 'Ova calificado',
+                icon: 'success',
+                confirmButtonText: 'aceptar',
+                width: '20%',
+                backdrop: false,
+                timer: 3000,
+                toast: true,
+                position: 'bottom-end'
+            })
             return this.http.post<any>(this.baseUrl + 'merge', newScore);
+
         }
 
     }
